xhanged type of amount in user table

+++ schemas.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import DeclarativeBase
 from sqlalchemy.orm import Mapped
 from sqlalchemy.orm import mapped_column
-from sqlalchemy.types import String, Integer
+from sqlalchemy.types import String, Float, Integer
 from sqlalchemy import create_engine
 from dotenv import dotenv_values
 
@@ -15,7 +15,7 @@ class Base(DeclarativeBase):
 class User(Base):
     __tablename__="user"
     tg_id: Mapped[int] = mapped_column(Integer())
-    amount: Mapped[int] = mapped_column(Integer())
+    amount: Mapped[float] = mapped_column(Float())
 
 if __name__ == "__main__":
     Base.metadata.create_all(engine)
